feat: DELETE /user sql DELETE문

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -71,9 +71,14 @@ public class UserController {
 
 	@DeleteMapping("/user")
 	public void deleteUser(@RequestParam String name) {
+		String readSql = "SELECT * FROM user WHERE name = ?";
+		boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+		if(isUserNotExist) {
+			throw new IllegalArgumentException();
+		}
+
 		String sql = "DELETE FROM user WHERE name = ?";
 		jdbcTemplate.update(sql, name);
 	}
-
 
 }
